% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adicionar_bioma.R
\name{adicionar_bioma}
\alias{adicionar_bioma}
\title{Funçao para adicionar bioma aos dados do Monitora}
\usage{
adicionar_bioma(myData)
}
\arguments{
\item{nome_arquivo}{Nome do arquivo de dados}
}
\value{
myData com bioma
}
\description{
Funçao para adicionar bioma aos dados do Monitora
}
\examples{
adicionar_bioma()
}
\author{
Elildo Carvalho Jr
}
