% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atribuir_spp.R
\name{atribuir_spp}
\alias{atribuir_spp}
\title{Atribuir epiteto especifico com base na lista de spp previamente registradas na UC}
\usage{
atribuir_spp(myData)
}
\value{
especies atribuidas
}
\description{
Atribuir epiteto especifico com base na lista de spp previamente registradas na UC
}
\examples{
atribuir_spp()
}
\author{
Jorge Mario Herrera Lopera
}
