% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atualizar_categorias.R
\name{atualizar_categorias}
\alias{atualizar_categorias}
\title{atualizar as categorias taxonômicas superiores a gênero}
\usage{
atualizar_categorias(myData)
}
\value{
resultado do modelo de espaço de estados
}
\description{
atualizar as categorias taxonômicas superiores a gênero
}
\examples{
atualizar_categorias()
}
\author{
Jorge Mario Herrera Lopera
}
