% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atualizar_especies.R
\name{atualizar_especies}
\alias{atualizar_especies}
\title{Atualizar a taxonomia das espécies caso sejam sinônimos}
\usage{
atualizar_especies(myData)
}
\value{
resultado do modelo de espaço de estados
}
\description{
Atualizar a taxonomia das espécies caso sejam sinônimos
}
\examples{
atualizar_especies()
}
\author{
Jorge Mario Herrera Lopera
}
