% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_G.R
\name{calc_G}
\alias{calc_G}
\title{Funçao para estimar media geometrica}
\usage{
calc_G(x)
}
\value{
Estimativas da media geometrica das abundâncias relativas usando bootstrap
}
\description{
Funçao para estimar media geometrica
}
\examples{
calc_G()

}
\author{
Elildo Carvalho Jr
}
