% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_tx_encontro.R
\name{calc_tx_encontro}
\alias{calc_tx_encontro}
\title{Funçao para calcular taxa de encontro por visita}
\usage{
calc_tx_encontro(monitora, pop)
}
\arguments{
\item{pop}{Populacao alvo da função}
}
\value{
tibble com taxas de encontro por visita
}
\description{
Funçao para calcular taxa de encontro por visita
}
\examples{
calc_tx_encontro()
}
\author{
Elildo Carvalho Jr
}
