% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checar_taxonomia.R
\name{checar_taxonomia}
\alias{checar_taxonomia}
\title{Funçao para checar visualmente taxonomia dos dados do Monitora}
\usage{
checar_taxonomia(myData)
}
\arguments{
\item{myData}{Nome do objeto. Default myData}
}
\value{
tabela com taxononia
}
\description{
Funçao para checar visualmente taxonomia dos dados do Monitora
}
\examples{
checar_taxonomia()

}
