% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrigir_nome_uc.R
\name{corrigir_nome_uc}
\alias{corrigir_nome_uc}
\title{Funçao para corrigir nomes das UCs}
\usage{
corrigir_nome_uc(myData)
}
\arguments{
\item{myData}{Nome do arquivo de myData}
}
\value{
myData com nome de UCs corrigidos
}
\description{
Funçao para corrigir nomes das UCs
}
\examples{
corrigir_nome_uc()
}
\author{
Jorge Mario Herrera Lopera
}
