% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrigir_taxonomia_deprecated.R
\name{corrigir_taxonomia_deprecated}
\alias{corrigir_taxonomia_deprecated}
\title{Funçao para limpar dados do Monitora}
\usage{
corrigir_taxonomia_deprecated(myData)
}
\arguments{
\item{myData}{Nome do arquivo de myData}
}
\value{
myData corrigidos e limpos
}
\description{
Funçao para limpar dados do Monitora
}
\examples{
corrigir_taxonomia()
}
\author{
Elildo Carvalho Jr
}
