% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrigir_tipografia.R
\name{corrigir_tipografia}
\alias{corrigir_tipografia}
\title{Corrigir a tipografia dos níveis taxonômicos}
\usage{
corrigir_tipografia(myData)
}
\value{
resultado do modelo de espaço de estados
}
\description{
Corrigir a tipografia dos níveis taxonômicos
}
\examples{
corrigir_tipografia()
}
\author{
Jorge Mario Herrera Lopera
}
