% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criar_template.R
\name{criar_template}
\alias{criar_template}
\title{Funçao para criar template com NAs para anos sem dados}
\usage{
criar_template(monitora, pop)
}
\arguments{
\item{pop}{Populacao alvo da função}
}
\value{
template indicando anos sem coleta de dados
}
\description{
Funçao para criar template com NAs para anos sem dados
}
\examples{
criar_template()

}
