% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cruzar_taxon_uc.R
\name{cruzar_taxon_uc}
\alias{cruzar_taxon_uc}
\title{Funçao para limpar dados do Monitora}
\usage{
cruzar_taxon_uc(myData)
}
\arguments{
\item{myData}{Nome do arquivo de myData}
}
\value{
myData corrigidos e limpos
}
\description{
Funçao para limpar dados do Monitora
}
\examples{
cruzar_taxon_uc()
}
\author{
Elildo Carvalho Jr
}
