% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grafico_pizza.R
\name{grafico_pizza}
\alias{grafico_pizza}
\title{Funçao para rodar modelo SSM no jags}
\usage{
grafico_pizza()
}
\value{
resultado do modelo de espaço de estados
}
\description{
Funçao para rodar modelo SSM no jags
}
\examples{
grafico_pizza()
}
