% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ler_dados.R
\name{ler_dados}
\alias{ler_dados}
\title{Funçao para limpar myData do Monitora}
\usage{
ler_dados(nome_arquivo = "Planilha MastoAves Consolidada ate 2022.csv")
}
\arguments{
\item{nome_arquivo}{Nome do arquivo de dados}
}
\value{
myData corrigidos e limpos
}
\description{
Funçao para limpar myData do Monitora
}
\examples{
ler_dados()
}
\author{
Elildo Carvalho Jr
}
