% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotar_cv.R
\name{plotar_cv}
\alias{plotar_cv}
\title{Funçao para plotar CV como funcao do esforco}
\usage{
plotar_cv()
}
\value{
Gráfico do CV da taxa de crescimento em função do esforço
}
\description{
Funçao para plotar CV como funcao do esforco
}
\examples{
plotar_cv()

}
\author{
Elildo Carvalho Jr
}
