% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotar_G.R
\name{plotar_G}
\alias{plotar_G}
\title{Funçao para plotar media geometrica}
\usage{
plotar_G()
}
\value{
Gráfico da vriação da media geometrica ao longo do tempo
}
\description{
Funçao para plotar media geometrica
}
\examples{
plotar_G()

}
\author{
Elildo Carvalho Jr
}
