% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotar_lpi.R
\name{plotar_lpi}
\alias{plotar_lpi}
\title{Funçao para plotar media geometrica}
\usage{
plotar_lpi()
}
\value{
Gráfico do Living Planet Index para dados do ICMBio
}
\description{
Funçao para plotar media geometrica
}
\examples{
plotar_lpi()
}
\author{
Elildo Carvalho Jr
}
