% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotar_multipainel.R
\name{plotar_multipainel}
\alias{plotar_multipainel}
\title{Funçao para plotar tendencias temporais formato barras}
\usage{
plotar_multipainel(tipo = c("uc", "r"), nome_uc, n_colunas = 2)
}
\value{
grafico de tendencias temporais
}
\description{
Funçao para plotar tendencias temporais formato barras
}
\examples{
plotar_multipainel()
plotar_multipainel(tipo = "uc", nome_uc = "Resex Tapajós-Arapiuns")
plotar_multipainel(tipo = "r")
}
\author{
Elildo Carvalho Jr
}
