% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotar_r.R
\name{plotar_r}
\alias{plotar_r}
\title{Funçao para plotar taxa de crescimento r}
\usage{
plotar_r()
}
\value{
histograma da taxa de crescimento r
}
\description{
Funçao para plotar taxa de crescimento r
}
\examples{
plotar_r()

}
\author{
Elildo Carvalho Jr
}
