% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotar_ssm_barra.R
\name{plotar_ssm_barra}
\alias{plotar_ssm_barra}
\title{Funçao para plotar tendencias temporais formato barras}
\usage{
plotar_ssm_barra()
}
\value{
grafico de tendencias temporais
}
\description{
Funçao para plotar tendencias temporais formato barras
}
\examples{
plotar_ssm_barra()
}
\author{
Elildo Carvalho Jr
}
