% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotar_ssm_linha.R
\name{plotar_ssm_linha}
\alias{plotar_ssm_linha}
\title{Funçao para plotar tendencias temporais formato linhas e ribbon}
\usage{
plotar_ssm_linha()
}
\value{
grafico de tendencias temporais
}
\description{
Funçao para plotar tendencias temporais formato linhas e ribbon
}
\examples{
plotar_ssm_linha()

}
\author{
Elildo Carvalho Jr
}
