% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rodar_ssm2.R
\name{rodar_ssm2}
\alias{rodar_ssm2}
\title{Funçao para rodar modelo SSM no jags}
\usage{
rodar_ssm2(monitora, pop)
}
\value{
resultado do modelo de espaço de estados
}
\description{
Funçao para rodar modelo SSM no jags
}
\examples{
rodar_ssm()
}
\author{
Elildo Carvalho Jr
}
