% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selecionar_pops_taxa_min.R
\name{selecionar_pops_taxa_min}
\alias{selecionar_pops_taxa_min}
\title{funcao para selecionar populacoes com taxa de encontro medio > valor especificado}
\usage{
selecionar_pops_taxa_min(monitora, taxa_min = 0.5)
}
\arguments{
\item{taxa_min}{Taxa mínima aceitável. Default = 0.5}
}
\value{
Dados filtrados
}
\description{
funcao para selecionar populacoes com taxa de encontro medio > valor especificado
}
\examples{
selecionar_pops_taxa_min()

}
\author{
Elildo Carvalho Jr
}
