% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selecionar_uc_serie_min.R
\name{selecionar_uc_serie_min}
\alias{selecionar_uc_serie_min}
\title{função para selecionar UCS com serie temporal >= 5 anos}
\usage{
selecionar_uc_serie_min(monitora)
}
\arguments{
\item{monitora}{Nome da base de dados}
}
\value{
Dados filtrados
}
\description{
função para selecionar UCS com serie temporal >= 5 anos
}
\examples{
selecionar_uc_serie_min()

}
\author{
Elildo Carvalho Jr
}
