% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verificar_cv.R
\name{verificar_cv}
\alias{verificar_cv}
\title{Funçao para verificar CV das taxas de encontro usando bootstrap}
\usage{
verificar_cv(popsToUse = c(popsToUse))
}
\arguments{
\item{monitora}{Nome da base de dados}
}
\value{
tabela comparando CV e esforco para todas as populacoes
}
\description{
Funçao para verificar CV das taxas de encontro usando bootstrap
}
\examples{
verificar_cv()

}
\author{
Elildo Carvalho Jr
}
